+++ src/services/inclusaoDeAlimentacao/shared.service.js
@@ -1,5 +1,7 @@
 import { FLUXO, AUTH_TOKEN } from "services/constants";
+import { ErrorHandlerFunction } from "services/service-helpers";
 import { getPath } from "./helper";
+import axios from "../_base";
 
 export const obterSolicitacaoDeInclusaoDeAlimentacao = (
   uuid,
@@ -19,19 +21,13 @@ export const obterSolicitacaoDeInclusaoDeAlimentacao = (
     });
 };
 
-export const obterMinhasSolicitacoesDeInclusaoDeAlimentacao = tipoSolicitacao => {
+export const obterMinhasSolicitacoesDeInclusaoDeAlimentacao = async tipoSolicitacao => {
   const url = `${getPath(tipoSolicitacao)}/minhas-solicitacoes/`;
-  const OBJ_REQUEST = {
-    headers: AUTH_TOKEN,
-    method: "GET"
-  };
-  return fetch(url, OBJ_REQUEST)
-    .then(result => {
-      return result.json();
-    })
-    .catch(error => {
-      console.log(error);
-    });
+  const response = await axios.get(url).catch(ErrorHandlerFunction);
+  if (response) {
+    const data = { data: response.data, status: response.status };
+    return data;
+  }
 };
 
 export const inicioPedido = (uuid, tipoSolicitacao) => {
